pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["panwoo1 <hvcloc@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
django = "^4.2.7"
djangorestframework = "^3.14.0"
ctransformers = "^0.2.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
